LayoutTests/imported/w3c/web-platform-tests/css/css-values/tree-counting/calc-sibling-function-in-shadow-dom-expected.txt

Some element.
Some other element.
Some third element.

FAIL Host children have sibling-index() and sibling-count() based on assignedNodes order assert_equals: expected "3" but got "1"
